moved files around to fix the starting issues and added the index.js lines needed.
