+++ src/Components/main.js
@@ -3,7 +3,7 @@ import { Switch } from 'react-mdl';
 import LandingPage from './landingpage';
 import { Route } from 'react-router-dom';
 import Home from './home';
-import Contact from './contact'
+// import Contact from './contact'
 
 
 
@@ -11,7 +11,7 @@ const Main = () => (
 <Switch>
     <Route exact path= "/" component={LandingPage} />
     <Route path= "/home" component={Home} />
-    <Route path= "/contact" component={Contact} />
+    {/* <Route path= "/contact" component={Contact} /> */}
 
 </Switch>
 )
